+++ packages/manager/src/features/linodes/LinodesDetail/LinodeAdvanced/LinodeAdvancedConfigurationsPanel.tsx
@@ -1,115 +1,58 @@
-import { Disk } from '@linode/api-v4/lib/linodes';
 import * as React from 'react';
-import { Link } from 'react-router-dom';
 import { compose } from 'recompose';
-import Paper from 'src/components/core/Paper';
-import {
-  createStyles,
-  Theme,
-  withStyles,
-  WithStyles
-} from 'src/components/core/styles';
-import Typography from 'src/components/core/Typography';
+import { makeStyles, Theme } from 'src/components/core/styles';
 import Grid from 'src/components/Grid';
 import { withLinodeDetailContext } from 'src/features/linodes/LinodesDetail/linodeDetailContext';
 import LinodeConfigs from './LinodeConfigs';
-import LinodeDisks from './LinodeDisks';
-import LinodeDiskSpace from './LinodeDiskSpace';
 
-import { sendMigrationNavigationEvent } from 'src/utilities/ga';
-
-type ClassNames =
-  | 'root'
-  | 'title'
-  | 'paper'
-  | 'migrationHeader'
-  | 'migrationCopy'
-  | 'sidebar';
-
-const styles = (theme: Theme) =>
-  createStyles({
-    title: {
-      marginBottom: theme.spacing(2)
+const useStyles = makeStyles((theme: Theme) => ({
+  title: {
+    marginBottom: theme.spacing(2),
+  },
+  enclosingGrid: {
+    width: '100%',
+  },
+  paper: {
+    padding: theme.spacing(3),
+    paddingTop: theme.spacing(1),
+    marginBottom: theme.spacing(3),
+  },
+  migrationHeader: {
+    paddingTop: theme.spacing(),
+  },
+  migrationCopy: {
+    marginTop: theme.spacing(),
+  },
+  sidebar: {
+    marginTop: -theme.spacing(2),
+    [theme.breakpoints.up('md')]: {
+      marginTop: theme.spacing(2) + 24,
     },
-    paper: {
-      padding: theme.spacing(3),
-      paddingTop: theme.spacing(1),
-      marginBottom: theme.spacing(3)
-    },
-    migrationHeader: {
-      paddingTop: theme.spacing()
-    },
-    migrationCopy: {
-      marginTop: theme.spacing()
-    },
-    sidebar: {
-      marginTop: -theme.spacing(2),
-      [theme.breakpoints.up('md')]: {
-        marginTop: theme.spacing(2) + 24
-      }
-    }
-  });
+  },
+}));
 
-type CombinedProps = LinodeContextProps & WithStyles<ClassNames>;
+type CombinedProps = LinodeContextProps;
 
-class LinodeAdvancedConfigurationsPanel extends React.PureComponent<
-  CombinedProps
-> {
-  render() {
-    const { classes, disks, linodeTotalDisk, linodeID } = this.props;
+const LinodeAdvancedConfigurationsPanel: React.FC<CombinedProps> = () => {
+  const classes = useStyles();
 
-    return (
-      <Grid container>
-        <Grid item xs={12} md={7} lg={9}>
-          <Typography variant="h2" className={classes.title}>
-            Linode Disks and Configurations
-          </Typography>
-          <Paper className={classes.paper}>
-            <LinodeConfigs />
-          </Paper>
-          <Paper className={classes.paper}>
-            <LinodeDisks />
-          </Paper>
-          <Paper className={classes.paper}>
-            <Typography variant="h3" className={classes.migrationHeader}>
-              Configure a Migration
-            </Typography>
-            <Typography className={classes.migrationCopy}>
-              You can migrate your Linode across datacenters automatically.
-              <Link
-                to={`/linodes/${linodeID}/migrate`}
-                onClick={() => sendMigrationNavigationEvent(`/advanced`)}
-              >
-                {' '}
-                Click here to get started.
-              </Link>
-            </Typography>
-          </Paper>
-        </Grid>
-        <Grid item xs={12} md={5} lg={3} className={classes.sidebar}>
-          <Paper className={classes.paper}>
-            <LinodeDiskSpace disks={disks} totalDiskSpace={linodeTotalDisk} />
-          </Paper>
-        </Grid>
+  return (
+    <Grid container item xs={12}>
+      <Grid item className={classes.enclosingGrid}>
+        <LinodeConfigs />
       </Grid>
-    );
-  }
-}
+    </Grid>
+  );
+};
 
 interface LinodeContextProps {
-  linodeTotalDisk: number;
-  disks: Disk[];
   linodeID: number;
 }
 
 const linodeContext = withLinodeDetailContext(({ linode }) => ({
-  linodeTotalDisk: linode.specs.disk,
-  disks: linode._disks,
-  linodeID: linode.id
+  linodeID: linode.id,
 }));
 
-const styled = withStyles(styles);
-
-const enhanced = compose<CombinedProps, {}>(styled, linodeContext);
+const enhanced = compose<CombinedProps, {}>(linodeContext);
 
 export default enhanced(LinodeAdvancedConfigurationsPanel);
