+++ packages/manager/src/features/linodes/LinodesDetail/LinodeAdvanced/LinodeConfigs.tsx
@@ -1,11 +1,23 @@
-import { Config, linodeReboot } from '@linode/api-v4/lib/linodes';
+import { Account } from '@linode/api-v4/lib/account';
+import {
+  Config,
+  Disk,
+  getInterfaces,
+  getLinodeKernels,
+  Kernel,
+  LinodeInterface,
+  linodeReboot,
+} from '@linode/api-v4/lib/linodes';
 import { APIError } from '@linode/api-v4/lib/types';
+import { VLAN } from '@linode/api-v4/lib/vlans';
+import { Volume } from '@linode/api-v4/lib/volumes';
 import { withSnackbar, WithSnackbarProps } from 'notistack';
 import * as React from 'react';
 import { connect } from 'react-redux';
 import { compose } from 'recompose';
+import { bindActionCreators, Dispatch } from 'redux';
 import ActionsPanel from 'src/components/ActionsPanel';
-import AddNewLink from 'src/components/AddNewLink';
+import AddNewLink from 'src/components/AddNewLink/AddNewLink_CMR';
 import Button from 'src/components/Button';
 import ConfirmationDialog from 'src/components/ConfirmationDialog';
 import RootRef from 'src/components/core/RootRef';
@@ -13,62 +25,120 @@ import {
   createStyles,
   Theme,
   withStyles,
-  WithStyles
+  WithStyles,
 } from 'src/components/core/styles';
 import TableBody from 'src/components/core/TableBody';
-import TableCell from 'src/components/core/TableCell';
 import TableHead from 'src/components/core/TableHead';
 import TableRow from 'src/components/core/TableRow';
 import Typography from 'src/components/core/Typography';
 import Grid from 'src/components/Grid';
+import OrderBy from 'src/components/OrderBy';
+import Paginate from 'src/components/Paginate';
 import PaginationFooter from 'src/components/PaginationFooter';
 import PanelErrorBoundary from 'src/components/PanelErrorBoundary';
-import Table from 'src/components/Table';
+import Table from 'src/components/Table/Table_CMR';
+import TableCell from 'src/components/TableCell/TableCell_CMR.tsx';
 import TableContentWrapper from 'src/components/TableContentWrapper';
+import TableSortCell from 'src/components/TableSortCell/TableSortCell_CMR';
+import withFeatureFlags, {
+  FeatureFlagConsumerProps,
+} from 'src/containers/withFeatureFlagConsumer.container.ts';
 import { resetEventsPolling } from 'src/eventsPolling';
 import {
   DeleteLinodeConfig,
-  withLinodeDetailContext
+  withLinodeDetailContext,
 } from 'src/features/linodes/LinodesDetail/linodeDetailContext';
 import { MapState } from 'src/store/types';
+import { getAllVlans } from 'src/store/vlans/vlans.requests';
+import { isFeatureEnabled } from 'src/utilities/accountCapabilities';
 import { getAPIErrorOrDefault } from 'src/utilities/errorUtils';
+import { getAll, GetAllData } from 'src/utilities/getAll';
 import LinodeConfigDrawer from '../LinodeSettings/LinodeConfigDrawer';
 import ConfigRow from './ConfigRow';
 
-import Paginate from 'src/components/Paginate';
-
-type ClassNames = 'root' | 'headline' | 'addNewWrapper';
+type ClassNames =
+  | 'root'
+  | 'gridContainer'
+  | 'headline'
+  | 'addNewWrapper'
+  | 'labelCell'
+  | 'tableCell'
+  | 'labelColumn'
+  | 'vmColumn'
+  | 'memoryColumn'
+  | 'kernelColumn'
+  | 'interfacesColumn'
+  | 'deviceColumn'
+  | 'actionsColumn';
 
 const styles = (theme: Theme) =>
   createStyles({
-    root: {},
+    root: {
+      backgroundColor: theme.color.white,
+      margin: 0,
+      width: '100%',
+    },
     headline: {
-      marginBottom: theme.spacing(2),
-      [theme.breakpoints.down('xs')]: {
-        marginBottom: 0,
-        marginTop: theme.spacing(2)
-      }
+      marginTop: 8,
+      marginBottom: 8,
+      marginLeft: 15,
+      lineHeight: '1.5rem',
     },
     addNewWrapper: {
       [theme.breakpoints.down('xs')]: {
-        width: '100%',
         marginLeft: -(theme.spacing(1) + theme.spacing(1) / 2),
-        marginTop: -theme.spacing(1)
+        marginTop: -theme.spacing(1),
       },
-      paddingRight: 5
-    }
+      padding: '5px !important',
+    },
+    labelCell: {
+      width: '25%',
+    },
+    tableCell: {
+      borderRight: `1px solid ${theme.palette.divider}`,
+      fontWeight: 'bold',
+    },
+    labelColumn: {
+      ...theme.applyTableHeaderStyles,
+      width: '18%',
+    },
+    vmColumn: {
+      ...theme.applyTableHeaderStyles,
+      width: '10%',
+    },
+    kernelColumn: {
+      width: '23%',
+    },
+    interfacesColumn: {
+      width: '20%',
+    },
+    deviceColumn: {
+      width: '20%',
+    },
+    actionsColumn: {
+      width: '10%',
+    },
   });
 
 type CombinedProps = LinodeContext &
+  FeatureFlagConsumerProps &
   WithStyles<ClassNames> &
   WithSnackbarProps &
-  StateProps;
+  StateProps &
+  DispatchProps;
 
 interface State {
   configDrawer: ConfigDrawerState;
   confirmDelete: ConfirmDeleteState;
+  confirmBoot: ConfirmBootState;
   submitting: boolean;
   success?: string;
+  kernels: Kernel[];
+  kernelError: APIError[] | null;
+  kernelsLoading: boolean;
+  interfaces: LinodeInterface[];
+  interfacesLoading: boolean;
+  interfacesError: APIError[] | null;
 }
 
 interface ConfigDrawerState {
@@ -84,19 +154,101 @@ interface ConfirmDeleteState {
   error?: string;
 }
 
+interface ConfirmBootState {
+  open: boolean;
+  submitting: boolean;
+  configId?: number;
+  label?: string;
+  error?: string;
+}
+
+const getAllKernels = getAll<Kernel>(getLinodeKernels);
+
 class LinodeConfigs extends React.Component<CombinedProps, State> {
   defaultConfigDrawerState: ConfigDrawerState = {
-    open: false
+    open: false,
   };
 
   state: State = {
     submitting: false,
     confirmDelete: {
       open: false,
-      submitting: false
+      submitting: false,
     },
-    configDrawer: this.defaultConfigDrawerState
+    confirmBoot: {
+      open: false,
+      submitting: false,
+    },
+    configDrawer: this.defaultConfigDrawerState,
+    kernels: [],
+    kernelError: null,
+    kernelsLoading: false,
+    interfaces: [],
+    interfacesLoading: false,
+    interfacesError: null,
   };
+
+  requestKernels = (linodeHypervisor: 'kvm' | 'xen') => {
+    this.setState({ kernelsLoading: true, kernelError: null });
+
+    return getAllKernels({}, { [linodeHypervisor]: true })
+      .then(({ data: kernels }) => {
+        this.setState({
+          kernels,
+          kernelsLoading: false,
+        });
+      })
+      .catch((error) => {
+        this.setState({
+          kernelError: getAPIErrorOrDefault(error, 'Unable to load kernels.'),
+          kernelsLoading: false,
+        });
+      });
+  };
+
+  requestInterfaces = () => {
+    this.setState({ interfacesLoading: true, interfacesError: null });
+
+    getInterfaces(this.props.linodeId)
+      .then(({ data: interfaces }) => {
+        this.setState({
+          interfaces,
+          interfacesLoading: false,
+        });
+      })
+      .catch((error) => {
+        this.setState({
+          interfacesLoading: false,
+          interfacesError: getAPIErrorOrDefault(
+            error,
+            'Unable to load Linode Interfaces.'
+          ),
+        });
+      });
+  };
+
+  maybeRequestVlans = () => {
+    if (this.props.vlansLastUpdated === 0 && !this.props.vlansLoading) {
+      this.props.getAllVlans().catch(() => null); // Handle errors with Redux.
+    }
+  };
+
+  vlansEnabled = () => {
+    return isFeatureEnabled(
+      'Vlans',
+      Boolean(this.props.flags.vlans),
+      this.props.accountData?.capabilities ?? []
+    );
+  };
+
+  componentDidMount() {
+    this.requestKernels(this.props.linodeHypervisor);
+
+    if (this.vlansEnabled()) {
+      this.requestInterfaces();
+      this.maybeRequestVlans();
+    }
+  }
 
   configsPanel = React.createRef();
 
@@ -105,11 +257,16 @@ class LinodeConfigs extends React.Component<CombinedProps, State> {
 
     return (
       <React.Fragment>
-        <Grid container justify="space-between" alignItems="flex-end">
+        <Grid
+          container
+          justify="space-between"
+          alignItems="flex-end"
+          className={classes.root}
+        >
           <RootRef rootRef={this.configsPanel}>
-            <Grid item>
+            <Grid item className="p0">
               <Typography variant="h3" className={classes.headline}>
-                Configuration
+                Configurations
               </Typography>
             </Grid>
           </RootRef>
@@ -129,6 +286,9 @@ class LinodeConfigs extends React.Component<CombinedProps, State> {
           maxMemory={this.props.linodeMemory}
           onClose={this.resetConfigDrawer}
           open={this.state.configDrawer.open}
+          kernels={this.state.kernels}
+          kernelError={this.state.kernelError}
+          kernelsLoading={this.state.kernelsLoading}
         />
         <ConfirmationDialog
           title="Confirm Delete"
@@ -138,7 +298,20 @@ class LinodeConfigs extends React.Component<CombinedProps, State> {
           actions={this.deleteConfigConfirmationActions}
         >
           <Typography>
-            Are you sure you want to delete "{this.state.confirmDelete.label}?"
+            Are you sure you want to delete &quot;
+            {this.state.confirmDelete.label}?&quot;
+          </Typography>
+        </ConfirmationDialog>
+        <ConfirmationDialog
+          title="Confirm Boot"
+          error={this.state.confirmBoot.error}
+          open={this.state.confirmBoot.open}
+          onClose={this.cancelBoot}
+          actions={this.bootConfigConfirmationActions}
+        >
+          <Typography>
+            Are you sure you want to boot &quot;{this.state.confirmBoot.label}
+            ?&quot;
           </Typography>
         </ConfirmationDialog>
       </React.Fragment>
@@ -164,12 +337,33 @@ class LinodeConfigs extends React.Component<CombinedProps, State> {
     </ActionsPanel>
   );
 
+  bootConfigConfirmationActions = () => (
+    <ActionsPanel style={{ padding: 0 }}>
+      <Button
+        onClick={this.cancelBoot}
+        buttonType="cancel"
+        data-qa-cancel-delete
+      >
+        Cancel
+      </Button>
+      <Button
+        buttonType="primary"
+        destructive
+        loading={this.state.confirmBoot.submitting}
+        onClick={this.handleBoot}
+        data-qa-confirm-reboot
+      >
+        Boot
+      </Button>
+    </ActionsPanel>
+  );
+
   resetConfirmConfigDelete = () =>
     this.setConfirmDelete({
       open: false,
       id: undefined,
       submitting: false,
-      error: undefined
+      error: undefined,
     });
 
   setConfigDrawer = (obj: Partial<ConfigDrawerState>, fn?: () => void) =>
@@ -177,8 +371,8 @@ class LinodeConfigs extends React.Component<CombinedProps, State> {
       {
         configDrawer: {
           ...this.state.configDrawer,
-          ...obj
-        }
+          ...obj,
+        },
       },
       () => {
         if (fn) {
@@ -192,8 +386,8 @@ class LinodeConfigs extends React.Component<CombinedProps, State> {
       {
         confirmDelete: {
           ...this.state.confirmDelete,
-          ...obj
-        }
+          ...obj,
+        },
       },
       () => {
         if (fn) {
@@ -205,14 +399,14 @@ class LinodeConfigs extends React.Component<CombinedProps, State> {
   openConfigDrawerForCreation = () => {
     this.setConfigDrawer({
       open: true,
-      linodeConfigId: undefined
+      linodeConfigId: undefined,
     });
   };
 
   openForEditing = (config: Config) => {
     this.setConfigDrawer({
       open: true,
-      linodeConfigId: config.id
+      linodeConfigId: config.id,
     });
   };
 
@@ -220,20 +414,51 @@ class LinodeConfigs extends React.Component<CombinedProps, State> {
     this.setConfirmDelete({ open: true, id, label, error: undefined });
   };
 
-  handleBoot = (linodeId: number, configId: number, label: string) => {
-    linodeReboot(linodeId, configId)
+  confirmBoot = (configId: number, label: string) => {
+    this.setState({
+      confirmBoot: {
+        ...this.state.confirmBoot,
+        open: true,
+        error: undefined,
+        configId,
+        label,
+      },
+    });
+  };
+
+  cancelBoot = () => {
+    this.setState({ confirmBoot: { ...this.state.confirmBoot, open: false } });
+  };
+
+  handleBoot = () => {
+    this.setState({
+      confirmBoot: {
+        ...this.state.confirmBoot,
+        error: undefined,
+        submitting: true,
+      },
+    });
+
+    linodeReboot(this.props.linodeId, this.state.confirmBoot.configId)
       .then(() => {
+        this.setState({
+          confirmBoot: {
+            ...this.state.confirmBoot,
+            open: false,
+            submitting: false,
+          },
+        });
+
         resetEventsPolling();
       })
-      .catch(errorResponse => {
-        const errors = getAPIErrorOrDefault(
+      .catch((errorResponse) => {
+        const error = getAPIErrorOrDefault(
           errorResponse,
-          `Error booting ${label}`
-        );
-        errors.forEach((error: APIError) => {
-          this.props.enqueueSnackbar(error.reason, {
-            variant: 'error'
-          });
+          `Error booting ${this.state.confirmBoot.label}`
+        )[0].reason;
+
+        this.setState({
+          confirmBoot: { ...this.state.confirmBoot, submitting: false, error },
         });
       });
   };
@@ -241,7 +466,7 @@ class LinodeConfigs extends React.Component<CombinedProps, State> {
   deleteConfig = () => {
     this.setConfirmDelete({ submitting: true, error: undefined });
     const {
-      confirmDelete: { id: configId }
+      confirmDelete: { id: configId },
     } = this.state;
     if (!configId) {
       return;
@@ -252,92 +477,152 @@ class LinodeConfigs extends React.Component<CombinedProps, State> {
       .then(() => {
         this.setConfirmDelete(
           {
-            submitting: true
+            submitting: true,
           },
           () => {
             this.setConfirmDelete({
               submitting: false,
               open: false,
-              id: undefined
+              id: undefined,
             });
           }
         );
       })
-      .catch(err => {
+      .catch((err) => {
         this.setConfirmDelete({
           submitting: false,
           error: getAPIErrorOrDefault(err, 'Unable to delete configuration.')[0]
-            .reason
+            .reason,
         });
       });
   };
 
   linodeConfigsTable = () => {
+    const { classes } = this.props;
+
     return (
-      <Paginate data={this.props.configs} scrollToRef={this.configsPanel}>
-        {({
-          data: paginatedData,
-          handlePageChange,
-          handlePageSizeChange,
-          page,
-          pageSize,
-          count
-        }) => {
-          const {
-            configsLastUpdated,
-            configsLoading,
-            configsError,
-            linodeId,
-            readOnly
-          } = this.props;
-          return (
-            <React.Fragment>
-              <Table
-                isResponsive={false}
-                aria-label="List of Configurations"
-                border
-              >
-                <TableHead>
-                  <TableRow>
-                    <TableCell>Label</TableCell>
-                    <TableCell />
-                  </TableRow>
-                </TableHead>
-                <TableBody>
-                  <TableContentWrapper
-                    loading={configsLoading && configsLastUpdated === 0}
-                    lastUpdated={configsLastUpdated}
-                    length={paginatedData.length}
-                    error={configsError}
-                  >
-                    {paginatedData.map(thisConfig => {
-                      return (
-                        <ConfigRow
-                          key={`config-row-${thisConfig.id}`}
-                          config={thisConfig}
-                          linodeId={linodeId}
-                          onBoot={this.handleBoot}
-                          onEdit={this.openForEditing}
-                          onDelete={this.confirmDelete}
-                          readOnly={readOnly}
-                        />
-                      );
-                    })}
-                  </TableContentWrapper>
-                </TableBody>
-              </Table>
-              <PaginationFooter
-                count={count}
-                page={page}
-                pageSize={pageSize}
-                handlePageChange={handlePageChange}
-                handleSizeChange={handlePageSizeChange}
-                eventCategory="linode configs"
-              />
-            </React.Fragment>
-          );
-        }}
-      </Paginate>
+      <OrderBy data={this.props.configs} orderBy={'label'} order={'asc'}>
+        {({ data: orderedData, handleOrderChange, order, orderBy }) => (
+          <Paginate data={orderedData} scrollToRef={this.configsPanel}>
+            {({
+              data: paginatedData,
+              handlePageChange,
+              handlePageSizeChange,
+              page,
+              pageSize,
+              count,
+            }) => {
+              const {
+                configsLastUpdated,
+                configsLoading,
+                configsError,
+                linodeId,
+                readOnly,
+              } = this.props;
+              return (
+                <React.Fragment>
+                  <Table aria-label="List of Configurations" border>
+                    <TableHead>
+                      <TableRow>
+                        <TableSortCell
+                          active={orderBy === 'label'}
+                          label={'label'}
+                          direction={order}
+                          handleClick={handleOrderChange}
+                          data-qa-config-label-header
+                          className={classes.labelColumn}
+                        >
+                          <strong>Label</strong>
+                        </TableSortCell>
+                        <TableSortCell
+                          active={orderBy === 'virt_mode'}
+                          label={'virt_mode'}
+                          direction={order}
+                          handleClick={handleOrderChange}
+                          data-qa-virt-mode-header
+                          className={classes.vmColumn}
+                        >
+                          <strong>VM Mode</strong>
+                        </TableSortCell>
+                        <TableCell
+                          className={`${classes.tableCell} ${classes.kernelColumn}`}
+                        >
+                          Kernel
+                        </TableCell>
+
+                        <TableCell
+                          className={`${classes.tableCell} ${classes.deviceColumn}`}
+                        >
+                          Disks
+                        </TableCell>
+                        {this.vlansEnabled() ? (
+                          <TableCell
+                            className={`${classes.tableCell} ${classes.interfacesColumn}`}
+                          >
+                            Network Interfaces
+                          </TableCell>
+                        ) : null}
+                        <TableCell className={classes.actionsColumn} />
+                      </TableRow>
+                    </TableHead>
+                    <TableBody>
+                      <TableContentWrapper
+                        loading={
+                          (configsLoading && configsLastUpdated === 0) ||
+                          this.state.kernelsLoading ||
+                          this.state.interfacesLoading ||
+                          this.props.vlansLoading
+                        }
+                        lastUpdated={configsLastUpdated}
+                        length={paginatedData.length}
+                        error={
+                          configsError ??
+                          this.state.interfacesError ??
+                          this.props.vlansError ??
+                          undefined
+                        }
+                      >
+                        {paginatedData.map((thisConfig) => {
+                          const kernel = this.state.kernels.find(
+                            (kernelName) => kernelName.id === thisConfig.kernel
+                          );
+                          return (
+                            <ConfigRow
+                              key={`config-row-${thisConfig.id}`}
+                              config={thisConfig}
+                              linodeId={linodeId}
+                              linodeMemory={this.props.linodeMemory}
+                              linodeDisks={this.props.linodeDisks}
+                              linodeKernel={kernel?.label ?? thisConfig.kernel}
+                              linodeIPs={this.props.linodeIPs}
+                              onBoot={this.confirmBoot}
+                              onEdit={this.openForEditing}
+                              onDelete={this.confirmDelete}
+                              readOnly={readOnly}
+                              linodeInterfaces={this.state.interfaces}
+                              linodeVolumes={this.props.linodeVolumes}
+                              vlans={this.props.vlansData}
+                              vlansEnabled={this.vlansEnabled()}
+                            />
+                          );
+                        })}
+                      </TableContentWrapper>
+                    </TableBody>
+                  </Table>
+                  <PaginationFooter
+                    count={count}
+                    page={page}
+                    pageSize={pageSize}
+                    handlePageChange={handlePageChange}
+                    handleSizeChange={handlePageSizeChange}
+                    eventCategory="linode configs"
+                  />
+                </React.Fragment>
+              );
+            }}
+          </Paginate>
+        )}
+      </OrderBy>
     );
   };
 }
@@ -345,7 +630,7 @@ class LinodeConfigs extends React.Component<CombinedProps, State> {
 const styled = withStyles(styles);
 
 const errorBoundary = PanelErrorBoundary({
-  heading: 'Advanced Configurations'
+  heading: 'Advanced Configurations',
 });
 
 interface LinodeContext {
@@ -355,11 +640,14 @@ interface LinodeContext {
   linodeMemory: number;
   linodeRegion: string;
   linodeStatus: string;
+  linodeIPs: string[];
   linodeTotalDisk: number;
   deleteLinodeConfig: DeleteLinodeConfig;
   readOnly: boolean;
   configs: Config[];
   getLinodeConfigs: () => void;
+  linodeDisks: Disk[];
+  linodeVolumes: Volume[];
 }
 
 const linodeContext = withLinodeDetailContext<LinodeContext>(
@@ -371,10 +659,13 @@ const linodeContext = withLinodeDetailContext<LinodeContext>(
     linodeRegion: linode.region,
     linodeStatus: linode.status,
     linodeTotalDisk: linode.specs.disk,
+    linodeIPs: linode.ipv4,
     readOnly: linode._permissions === 'read_only',
     deleteLinodeConfig,
     configs: linode._configs,
-    getLinodeConfigs
+    getLinodeConfigs,
+    linodeDisks: linode._disks,
+    linodeVolumes: linode._volumes,
   })
 );
 
@@ -382,6 +673,11 @@ interface StateProps {
   configsError?: APIError[];
   configsLoading: boolean;
   configsLastUpdated: number;
+  vlansData: Record<string, VLAN>;
+  vlansLoading: boolean;
+  vlansLastUpdated: number;
+  vlansError?: APIError[];
+  accountData?: Account;
 }
 
 const mapStateToProps: MapState<StateProps, LinodeContext> = (
@@ -392,14 +688,32 @@ const mapStateToProps: MapState<StateProps, LinodeContext> = (
   return {
     configsLastUpdated: configState?.lastUpdated ?? 0,
     configsLoading: configState?.loading ?? false,
-    configsError: configState?.error.read ?? undefined
+    configsError: configState?.error.read ?? undefined,
+    vlansData: state.__resources.vlans.itemsById,
+    vlansLoading: state.__resources.vlans.loading,
+    vlansLastUpdated: state.__resources.vlans.lastUpdated,
+    vlansError: state.__resources.vlans.error.read,
+    accountData: state.__resources.account.data,
   };
 };
 
-const connected = connect(mapStateToProps);
+interface DispatchProps {
+  getAllVlans: () => Promise<GetAllData<VLAN>>;
+}
+
+const mapDispatchToProps = (dispatch: Dispatch) =>
+  bindActionCreators(
+    {
+      getAllVlans,
+    },
+    dispatch
+  );
+
+const connected = connect(mapStateToProps, mapDispatchToProps);
 
 const enhanced = compose<CombinedProps, {}>(
   linodeContext,
+  withFeatureFlags,
   connected,
   styled,
   errorBoundary,
