+++ packages/manager/src/features/linodes/LinodesDetail/LinodeSettings/LinodeConfigActionMenu.tsx
@@ -1,102 +1,123 @@
 import { Config } from '@linode/api-v4/lib/linodes';
+import { splitAt } from 'ramda';
 import * as React from 'react';
 import { RouteComponentProps, withRouter } from 'react-router-dom';
-
-import ActionMenu, { Action } from 'src/components/ActionMenu/ActionMenu';
+import ActionMenu, {
+  Action,
+} from 'src/components/ActionMenu_CMR/ActionMenu_CMR';
+import {
+  makeStyles,
+  Theme,
+  useMediaQuery,
+  useTheme,
+} from 'src/components/core/styles';
+import InlineMenuAction from 'src/components/InlineMenuAction';
 
 interface Props {
   onEdit: (config: Config) => void;
   onDelete: (id: number, label: string) => void;
-  onBoot: (linodeId: number, configId: number, label: string) => void;
+  onBoot: (configId: number, label: string) => void;
   config: Config;
   linodeId: number;
   readOnly?: boolean;
   label: string;
 }
 
+const useStyles = makeStyles(() => ({
+  root: {
+    display: 'flex',
+    justifyContent: 'flex-end',
+    alignItems: 'center',
+  },
+}));
+
 type CombinedProps = Props & RouteComponentProps<{}>;
 
-class ConfigActionMenu extends React.Component<CombinedProps> {
-  handleEdit = () => {
-    const { config, onEdit } = this.props;
-    onEdit(config);
-  };
+const ConfigActionMenu: React.FC<CombinedProps> = (props) => {
+  const { readOnly, history, linodeId, config } = props;
 
-  handleDelete = () => {
+  const classes = useStyles();
+  const theme = useTheme<Theme>();
+  const matchesSmDown = useMediaQuery(theme.breakpoints.down('sm'));
+
+  const handleEdit = React.useCallback(() => {
+    const { onEdit } = props;
+    onEdit(config);
+  }, [config, props]);
+
+  const handleBoot = React.useCallback(() => {
+    const { onBoot } = props;
+    const { id, label } = config;
+    onBoot(id, label);
+  }, [config, props]);
+
+  const tooltip = readOnly
+    ? "You don't have permission to perform this action"
+    : undefined;
+
+  const handleDelete = () => {
     const {
       config: { id, label },
-      onDelete
-    } = this.props;
+      onDelete,
+    } = props;
     onDelete(id, label);
   };
 
-  handleBoot = () => {
-    const {
-      config: { id, label },
-      linodeId,
-      onBoot
-    } = this.props;
-    onBoot(linodeId, id, label);
-  };
+  const actions: Action[] = [
+    {
+      title: 'Boot',
+      onClick: () => {
+        handleBoot();
+      },
+    },
+    {
+      title: 'Edit',
+      onClick: () => {
+        handleEdit();
+      },
+    },
+    {
+      title: 'Clone',
+      onClick: () => {
+        history.push(
+          `/linodes/${linodeId}/clone/configs?selectedConfig=${config.id}`
+        );
+      },
+      disabled: readOnly,
+    },
+    {
+      title: 'Delete',
+      onClick: () => {
+        handleDelete();
+      },
+      disabled: readOnly,
+      tooltip,
+    },
+  ];
 
-  createConfigActions = () => (closeMenu: Function): Action[] => {
-    const { readOnly, history, linodeId, config } = this.props;
-    const tooltip = readOnly
-      ? "You don't have permission to perform this action"
-      : undefined;
-    return [
-      {
-        title: 'Boot This Config',
-        onClick: (e: React.MouseEvent<HTMLElement>) => {
-          e.preventDefault();
-          this.handleBoot();
-          closeMenu();
-        },
-        disabled: readOnly,
-        tooltip
-      },
-      {
-        title: 'Edit',
-        onClick: (e: React.MouseEvent<HTMLElement>) => {
-          e.preventDefault();
-          this.handleEdit();
-          closeMenu();
-        },
-        disabled: readOnly,
-        tooltip
-      },
-      {
-        title: 'Clone',
-        onClick: (e: React.MouseEvent<HTMLElement>) => {
-          e.preventDefault();
-          closeMenu();
-          history.push(
-            `/linodes/${linodeId}/clone/configs?selectedConfig=${config.id}`
+  const splitActionsArrayIndex = matchesSmDown ? 0 : 2;
+  const [inlineActions, menuActions] = splitAt(splitActionsArrayIndex, actions);
+
+  return (
+    <div className={classes.root}>
+      {!matchesSmDown &&
+        inlineActions.map((action) => {
+          return (
+            <InlineMenuAction
+              key={action.title}
+              actionText={action.title}
+              onClick={action.onClick}
+              tooltip={action.tooltip}
+              disabled={action.disabled}
+            />
           );
-        },
-        disabled: readOnly
-      },
-      {
-        title: 'Delete',
-        onClick: (e: React.MouseEvent<HTMLElement>) => {
-          e.preventDefault();
-          this.handleDelete();
-          closeMenu();
-        },
-        disabled: readOnly,
-        tooltip
-      }
-    ];
-  };
-
-  render() {
-    return (
+        })}
       <ActionMenu
-        createActions={this.createConfigActions()}
-        ariaLabel={`Action menu for Linode Config ${this.props.label}`}
+        actionsList={menuActions}
+        ariaLabel={`Action menu for Linode Config ${props.label}`}
       />
-    );
-  }
-}
+    </div>
+  );
+};
 
 export default withRouter(ConfigActionMenu);
